user 약간 수정, 마이페이지 Header,Footer 넣었음, 마이페이지 판매내역 가져옴
