+++ backend/src/main/java/com/bitc/finalproject/entity/BookEntity.java
@@ -41,7 +41,4 @@ public class BookEntity {
     @Column(nullable = false)
     private String saleBookTitle;
 
-
-
-
 }
