+++ backend/src/main/java/com/bitc/finalproject/repository/BookInfoRepository.java
@@ -9,4 +9,7 @@ public interface BookInfoRepository  extends JpaRepository<BookEntity, Integer> 
     List<BookEntity> findAllByOrderBySalePkDesc();
 
     List<BookEntity> findBySaleBookId(String isbn13);
+
+    List<BookEntity> findBySaleSellerId(String userId);
+
 }
