+++ frontend/src/pages/member/Join.js
@@ -2,6 +2,8 @@ import React, {useEffect, useState} from 'react';
 import axios from "axios";
 import { useCookies } from 'react-cookie';
 import {Link, useNavigate} from "react-router-dom";
+import Header from "../mainPages/Header";
+import Footer from "../mainPages/Footer";
 
 function Join(props) {
 
@@ -56,6 +58,7 @@ function Join(props) {
 
     return (
         <div className={'container my-3'}>
+            <Header />
             <div className={'row'}>
                 <div className="col-sm-4 mx-auto">
                     {/*<form action="/login" method="post">*/}
@@ -79,6 +82,7 @@ function Join(props) {
                     {/*</form>*/}
                 </div>
             </div>
+            <Footer />
         </div>
     )
 }
